requirements.txt
pandas==0.24.2
PyYAML==5.3.1
pytest==4.6.9
